requirements.txt
absl-py==1.2.0
amqp==5.1.1
art==5.7
asgiref==3.5.2
aspose-words==24.4.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.1
backcall==0.2.0
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chefboost==0.0.17
ci-info==0.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clip==0.2.0
cmake==3.26.3
colorama==0.4.5
comm==0.1.3
configobj==5.0.8
configparser==5.3.0
contourpy==1.0.5
cryptography==39.0.0
cssbeautifier==1.14.7
cvxopt==1.3.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
dj-database-url==0.5.0
Django==4.1.2
django-allauth==0.52.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.1
django-private-storage==3.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
docx==0.2.4
docx2pdf==0.1.8
drf-spectacular==0.24.2
drf-spectacular-sidecar==2022.10.1
editdistance==0.6.0
EditorConfig==0.12.3
efficientnet==1.0.0
essential-generators==1.0
etelemetry==0.3.0
executing==1.2.0
filelock==3.12.0
flatbuffers==22.9.24
fonttools==4.37.4
future==0.18.3
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.49.1
gunicorn==20.1.0
h5py==3.7.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httplib2==0.22.0
idna==3.4
imageio==2.22.1
imgaug==0.4.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.22.0
ipython==8.11.0
isodate==0.6.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsbeautifier==1.14.7
jsonpickle==3.0.1
jsonschema==4.16.0
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.10.0
Keras-Applications==1.0.8
keras-ocr==0.8.9
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kombu==5.2.4
libclang==14.0.6
lit==16.0.2
looseversion==1.1.2
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==2.8.7
nibabel==5.1.0
nipype==1.8.6
numpy==1.23.3
numpyencoder==0.3.0
oauthlib==3.2.1
opencv-contrib-python==4.6.0.66
opencv-contrib-python-headless==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==2.2.2
parso==0.8.3
pathlib==1.0.1
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.4.0
platformdirs==3.2.0
plum-dispatch==1.7.4
pm4py==2.7.11.11
polars==1.8.2
prompt-toolkit==3.0.31
protobuf==3.19.6
prov==2.0.0
psutil==5.9.2
psycopg2==2.9.4
ptyprocess==0.7.0
PuLP==2.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post3
pycparser==2.21
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.14.0
pygraphviz==1.13
PyJWT==2.6.0
pypandoc==1.13
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.4
python-docx==0.8.11
python3-openid==3.2.0
pytz==2022.4
pyvis==0.3.2
PyWavelets==1.4.1
pyxnat==1.5
PyYAML==6.0
pyzmq==25.0.2
rdflib==6.3.2
redis==4.3.4
regex==2023.3.23
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sahi==0.11.15
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.0
semantic-version==2.10.0
shapely==2.0.2
simplejson==3.19.1
six==1.16.0
sklearn==0.0
sortedcontainers==2.4.0
sqlparse==0.4.3
stack-data==0.6.2
StringDist==1.0.9
sympy==1.11.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2022.10.10
toml==0.10.2
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tornado==6.3.1
tqdm==4.64.1
traitlets==5.9.0
traits==6.3.2
transformers==4.40.1
typing_extensions==4.4.0
tzdata==2024.1
uc-micro-py==1.0.3
ultralytics==8.2.74
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.12
validators==0.20.0
vine==5.0.0
wcwidth==0.2.5
websockets==11.0.3
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
yarl==1.9.4
